refactored home layout and cypress tests
